==> source/step2_data_collection.ipynb <==
### ❗IMPORTANT: Twitter / X has deprecated the API used in this notebook for free access. This notebook is no longer functional.
For more detailed information, please refer to the following links about the decision to deprecate the API:
[Twitter Post](https://twitter.com/XDevelopers/status/1649191521323995138)
[Twitter Community Discussion](https://twittercommunity.com/t/deprecation-of-legacy-access-tiers-in-the-twitter-api/196162)
and the following open letter:
[Open letter with the  Coalition for Independent Technology Research](https://independenttechresearch.org/letter-twitters-new-api-plans-will-devastate-public-interest-research)

You need to subscribe elevated or enterprise access to Twitter API to access tweets.

import pandas as pd
import tweepy

keywords_df = pd.read_csv('../data/intermediate/input/step_2_twitter_keywords.csv')
included_keywords = keywords_df[keywords_df['is_keyword'] == 1]['keyword'].tolist()
crawled_tweets = []

# Your app's bearer token can be found under the Authentication Tokens section
# of the Keys and Tokens tab of your app, under the
# Twitter Developer Portal Projects & Apps page at
# https://developer.twitter.com/en/portal/projects-and-apps
bearer_token = ""

# Your app's API/consumer key and secret can be found under the Consumer Keys
# section of the Keys and Tokens tab of your app, under the
# Twitter Developer Portal Projects & Apps page at
# https://developer.twitter.com/en/portal/projects-and-apps
consumer_key = ""
consumer_secret = ""

# Your account's (the app owner's account's) access token and secret for your
# app can be found under the Authentication Tokens section of the
# Keys and Tokens tab of your app, under the
# Twitter Developer Portal Projects & Apps page at
# https://developer.twitter.com/en/portal/projects-and-apps
access_token = ""
access_token_secret = ""

# You can authenticate as your app with just your bearer token
client = tweepy.Client(bearer_token=bearer_token)

# You can provide the consumer key and secret with the access token and access
# token secret to authenticate as a user
client = tweepy.Client(
    consumer_key=consumer_key, consumer_secret=consumer_secret,
    access_token=access_token, access_token_secret=access_token_secret
)

for keyword in included_keywords:
    print(f'Crawling tweets for keyword: {keyword}')
    tweets = client.search_recent_tweets(query=keyword, max_results=100)
    crawled_tweets.extend(tweets)

crawled_tweets_df = pd.DataFrame(crawled_tweets)
crawled_tweets_df.to_csv('../data/intermediate/output/step_2_crawled_tweets.csv', index=False)
